+++ db/migrate/20231226174925_add_long_term_loan_to_loans.rb
@@ -1,0 +1,5 @@
+class AddLongTermLoanToLoans < ActiveRecord::Migration[6.1]
+  def change
+    add_column :loans, :is_long_term_loan, :boolean
+  end
+end